# api/swagger.yaml
openapi: 3.0.0
info:
  title: Keep Charged
  description: API documentation
  contact: {}
  version: 0.0.1 beta
servers:
- url: /
...
paths:
...
  /api/station/{id}:
    get:
      tags:
      - Login required
      summary: Returns the specified users charging station
      parameters:
      - name: id
        in: path
        description: Station id
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "404":
          description: Specified action/object is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
    put:
      tags:
      - Login required
      summary: Creates new charging station instance
      operationId: mobile8
      parameters:
      - name: id
        in: path
        description: Configuration section name
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: section for updating
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/station_id_body'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "404":
          description: Specified action/object is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
        "422":
          description: Error appeared while parsing request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
    delete:
      tags:
      - Login required
      summary: Deletes the specified chargingstation
      operationId: mobile9
      parameters:
      - name: id
        in: path
        description: Operator lists section name
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        "404":
          description: Specified action/object is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_404'
...
components:
  schemas:
    default_error:
      title: default_error
      type: object
      properties:
        code:
          type: number
          description: Error code
        error:
          type: string
          description: Error message
        source:
          type: string
          description: Error source
      description: Default error object
...
    inline_response_404:
      type: object
      properties:
        success:
          type: boolean
          description: Operation status
          default: false
        errors:
          type: array
          items:
            $ref: '#/components/schemas/default_error'
...
    inline_response_200_2:
      type: object
      properties:
        success:
          type: boolean
          description: Operation status.
        data: {}
      example:
        data: ""
        success: true
    station_id_body:
      type: object
      properties:
        data: {}
    inline_response_200_3:
      type: object
      properties:
        success:
          type: boolean
          description: Operation status
      example:
        success: true
...